Upgrade database fields

+++ prisma/migrations/20230706033327_/migration.sql
@@ -6,8 +6,8 @@ CREATE TABLE "Review" (
     "createdAt" TIMESTAMP(3) NOT NULL DEFAULT CURRENT_TIMESTAMP,
     "updatedAt" TIMESTAMP(3) NOT NULL,
     "customer_name" TEXT NOT NULL,
-    "customer_id" INTEGER NOT NULL,
-    "product_id" INTEGER NOT NULL,
+    "customer_id" BIGINT NOT NULL,
+    "product_id" BIGINT NOT NULL,
 
     CONSTRAINT "Review_pkey" PRIMARY KEY ("id")
 );
